Moved SearchBar componets to their own folder and created and implement a Higher Order Component to consolidate all the code that was duplicated in the SearchBar and SearchBarMobile component.
